// src/api/models/DraftAssignment.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DraftAssignment
 */
export interface DraftAssignment {
    /**
     * 
     * @type {string}
     * @memberof DraftAssignment
     */
    playerName: string;
    /**
     * 
     * @type {number}
     * @memberof DraftAssignment
     */
    team: number;
    /**
     * 
     * @type {number}
     * @memberof DraftAssignment
     */
    positionNumber: number;
    /**
     * 
     * @type {number}
     * @memberof DraftAssignment
     */
    general: number;
}

/**
 * Check if a given object implements the DraftAssignment interface.
 */
export function instanceOfDraftAssignment(value: object): value is DraftAssignment {
    if (!('playerName' in value) || value['playerName'] === undefined) return false;
    if (!('team' in value) || value['team'] === undefined) return false;
    if (!('positionNumber' in value) || value['positionNumber'] === undefined) return false;
    if (!('general' in value) || value['general'] === undefined) return false;
    return true;
}

export function DraftAssignmentFromJSON(json: any): DraftAssignment {
    return DraftAssignmentFromJSONTyped(json, false);
}

export function DraftAssignmentFromJSONTyped(json: any, ignoreDiscriminator: boolean): DraftAssignment {
    if (json == null) {
        return json;
    }
    return {
        
        'playerName': json['playerName'],
        'team': json['team'],
        'positionNumber': json['positionNumber'],
        'general': json['general'],
    };
}

export function DraftAssignmentToJSON(json: any): DraftAssignment {
    return DraftAssignmentToJSONTyped(json, false);
}

export function DraftAssignmentToJSONTyped(value?: DraftAssignment | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'playerName': value['playerName'],
        'team': value['team'],
        'positionNumber': value['positionNumber'],
        'general': value['general'],
    };
}