// src/api/models/MapsByPlayerCount.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MapsByPlayerCount
 */
export interface MapsByPlayerCount {
    /**
     * 
     * @type {number}
     * @memberof MapsByPlayerCount
     */
    playerCount: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof MapsByPlayerCount
     */
    maps: Array<string>;
}

/**
 * Check if a given object implements the MapsByPlayerCount interface.
 */
export function instanceOfMapsByPlayerCount(value: object): value is MapsByPlayerCount {
    if (!('playerCount' in value) || value['playerCount'] === undefined) return false;
    if (!('maps' in value) || value['maps'] === undefined) return false;
    return true;
}

export function MapsByPlayerCountFromJSON(json: any): MapsByPlayerCount {
    return MapsByPlayerCountFromJSONTyped(json, false);
}

export function MapsByPlayerCountFromJSONTyped(json: any, ignoreDiscriminator: boolean): MapsByPlayerCount {
    if (json == null) {
        return json;
    }
    return {
        
        'playerCount': json['playerCount'],
        'maps': json['maps'],
    };
}

export function MapsByPlayerCountToJSON(json: any): MapsByPlayerCount {
    return MapsByPlayerCountToJSONTyped(json, false);
}

export function MapsByPlayerCountToJSONTyped(value?: MapsByPlayerCount | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'playerCount': value['playerCount'],
        'maps': value['maps'],
    };
}